`Added confirmation before deleting a contact and updated delete functionality`

+++ functions/deleteContact.php
@@ -2,8 +2,18 @@
     session_start();
     include "../config/db.php";
 
-    $sql = "DELETE FROM contacts WHERE id = ?";
-    $stmt = $conn -> prepare($sql);
-    $stmt -> execute([$_GET["contactId"]]);
-    header("Location: ../public/contacts.php?userId=" . urlencode($_SESSION["id"]));
+    if ($_SERVER["REQUEST_METHOD"] == "POST") {
+        if (empty($_POST["idContact"])) {
+            $errorMessage = "You can't leave an empty input";
+            header("Location: ../public/contacts.php?errorMessage=" . urlencode($errorMessage));
+        }else{
+            $id = htmlspecialchars($_POST["idContact"]);
+        }
+
+        $sql = "DELETE FROM contacts WHERE id = ?";
+        $stmt = $conn -> prepare($sql);
+        $stmt -> execute([$id]);
+        header("Location: ../public/contacts.php?userId=" . urlencode($_SESSION["id"]));
+    }
+
 ?>
